EEPROM fix and Boot animation

+++ src/core/LED.h
@@ -22,13 +22,13 @@ public:
   // void setLED(INDEXMoDE indexmode, LEDMoDE ledmode, u8 x1, u8 y1, u32 p1, u8 p2, u8 p3, u8 p4);
 
   //Raw index - for BottomLED ，PoWERCoRD and raw strap.
-  void off(u8 index, bool overlay = false);
-  void on(u8 index, bool overlay = false);
-  void setW(u8 index, u8 w, bool overlay = false);
-  void setRGB(u8 index, u8 r, u8 g, u8 b, bool overlay = false);
-  void setWRGB(u8 index, u8 w, u8 r, u8 g, u8 b, bool overlay = false);
-  void setHEX(u8 index, u32 hex, bool overlay = false, bool ignore_gamma = false);
-  void setPalette(u8 index, u8 pick_palette, u8 colour, bool overlay = false);
+  void off(s16 index, bool overlay = false);
+  void on(s16 index, bool overlay = false);
+  void setW(s16 index, u8 w, bool overlay = false);
+  void setRGB(s16 index, u8 r, u8 g, u8 b, bool overlay = false);
+  void setWRGB(s16 index, u8 w, u8 r, u8 g, u8 b, bool overlay = false);
+  void setHEX(s16 index, u32 hex, bool overlay = false, bool ignore_gamma = false);
+  void setPalette(s16 index, u8 pick_palette, u8 colour, bool overlay = false);
 
   //XY
   void offXY(u8 xy, bool overlay = false);
